Initial work to refactor and inline bluetooth stuff.
