+++ extra_protos/SocketInfoRequest.proto
@@ -1,6 +1,6 @@
 syntax = "proto2";
 
-package openauto.btservice.proto;
+package openauto.extraprotos;
 
 message SocketInfoRequest
 {
